Displays the Mandelbrot set(no zoom feature)
Initial commit of adding the react native app folder. The code up to this point can display the Mandelbrot set on the screen, the zoom feature is still yet to be added. The code works by generating multiple small rectangles(acting like a pixel on the screen) whose color is determined by the Mandelbrot set arithmetic.